Moved libs to root of the project
