+++ actions/getMessages.ts
@@ -1,4 +1,4 @@
-import prisma from "@/app/libs/prismadb";
+import prisma from "@/libs/prismadb";
 
 /**
  * Gets a list of messages for a given conversation.
