+++ app/api/auth/[...nextauth]/route.ts
@@ -5,7 +5,7 @@ import GithubProvider from "next-auth/providers/github";
 import GoogleProvider from "next-auth/providers/google";
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 
-import prisma from "@/app/libs/prismadb";
+import prisma from "@/libs/prismadb";
 
 /**
  * Object providing the configuration options for NextAuth, including the
