+++ app/api/conversations/[conversationId]/seen/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse } from "next/server";
 
 import getCurrentUser from "@/actions/getCurrentUser";
-import { pusherServer } from "@/app/libs/pusher";
-import prisma from "@/app/libs/prismadb";
+import { pusherServer } from "@/libs/pusher";
+import prisma from "@/libs/prismadb";
 
 interface IParams {
   conversationId?: string;
