+++ app/api/conversations/route.ts
@@ -1,8 +1,8 @@
 import getCurrentUser from "@/actions/getCurrentUser";
 import { NextResponse } from "next/server";
 
-import prisma from "@/app/libs/prismadb";
-import { pusherServer } from "@/app/libs/pusher";
+import prisma from "@/libs/prismadb";
+import { pusherServer } from "@/libs/pusher";
 
 /**
  * A post request route to create a new conversation.
