+++ app/api/register/route.ts
@@ -1,6 +1,6 @@
 import bcrypt from "bcrypt";
 
-import prisma from "@/app/libs/prismadb";
+import prisma from "@/libs/prismadb";
 import { NextResponse } from "next/server";
 
 /**
