+++ app/api/settings/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import getCurrentUser from "@/actions/getCurrentUser";
-import prisma from "@/app/libs/prismadb";
+import prisma from "@/libs/prismadb";
 
 /**
  * A post request route for updating the user's settings (name and image).
