+++ pages/api/pusher/auth.ts
@@ -2,7 +2,7 @@
 
 import { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth";
-import { pusherServer } from "@/app/libs/pusher";
+import { pusherServer } from "@/libs/pusher";
 import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 /**
